debug

+++ lab3-sw/vga_ball.c
@@ -86,7 +86,7 @@ static void write_coordinate(sv_map *data){
  */
 
 //this will write backgrounds for apple and snake sprites 
-static long vga_ball_ioctl(struct file *f, unsigned int cmd, uint32_t arg)
+static long vga_ball_ioctl(struct file *f, unsigned int cmd, unsigned long arg)
 {
 	vga_ball_arg_t vla;
